Unify encode_event implementations
